Mengganti Tipe data ke double

+++ 7_Branching - Looping - Function/praktikum/task_no2.dart
@@ -1,21 +1,21 @@
 void main() {
-  int nilaiA = 10;
-  int faktorialA = 1;
-  for (int i = 1; i <= nilaiA; i++) {
+  double nilaiA = 10;
+  double faktorialA = 1;
+  for (double i = 1; i <= nilaiA; i++) {
     faktorialA *= i;
   }
   print('Nilai Faktorial 10 = $faktorialA');
 
-  int nilaiB = 20;
-  int faktorialB = 1;
-  for (int i = 1; i <= nilaiB; i++) {
+  double nilaiB = 20;
+  double faktorialB = 1;
+  for (double i = 1; i <= nilaiB; i++) {
     faktorialB *= i;
   }
   print('Nilai Faktorial 20 = $faktorialB');
 
-  int nilaiC = 30;
-  int faktorialC = 1;
-  for (int i = 1; i <= nilaiC; i++) {
+  double nilaiC = 30;
+  double faktorialC = 1;
+  for (double i = 1; i <= nilaiC; i++) {
     faktorialC *= i;
   }
   print('Nilai Faktorial 30 = $faktorialC');
